[MOB-16085] -- Move models for recipes into SDK
