+++ core/hh-geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/GeminioSdk.kt
@@ -1,0 +1,15 @@
+package ru.hh.plugins.geminio.sdk
+
+import com.android.tools.idea.npw.template.ConfigureTemplateParametersStep
+
+
+/**
+ * SDK with Geminio functionality.
+ */
+interface GeminioSdk {
+
+    fun parseGeminioRecipe(
+        recipeFilePath: String
+    ): ConfigureTemplateParametersStep
+
+}
