+++ core/hh-geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/GeminioSdkFactory.kt
@@ -1,0 +1,10 @@
+package ru.hh.plugins.geminio.sdk
+
+
+object GeminioSdkFactory {
+
+    fun createGeminioSdk(): GeminioSdk {
+        TODO()
+    }
+
+}
