+++ core/hh-geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/model/enums/BuildGradleDependencyConfiguration.kt
@@ -1,4 +1,4 @@
-package ru.hh.plugins.model
+package ru.hh.plugins.geminio.sdk.model.enums
 
 
 /**
