+++ core/hh-geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/model/enums/GeminioRecipeExpressionModifier.kt
@@ -1,4 +1,4 @@
-package ru.hh.plugins.geminio.model.enums
+package ru.hh.plugins.geminio.sdk.model.enums
 
 
 enum class GeminioRecipeExpressionModifier(
