+++ core/hh-geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/model/recipe/RecipeExpression.kt
@@ -1,6 +1,6 @@
-package ru.hh.plugins.geminio.model
+package ru.hh.plugins.geminio.sdk.model.recipe
 
-import ru.hh.plugins.geminio.model.enums.GeminioRecipeExpressionModifier
+import ru.hh.plugins.geminio.sdk.model.enums.GeminioRecipeExpressionModifier
 
 
 data class RecipeExpression(
