+++ core/hh-plugins-core/build.gradle.kts
@@ -10,6 +10,8 @@ repositories {
 }
 
 dependencies {
+    implementation(project(":hh-geminio-sdk"))// todo remove
+
     implementation(kotlin("stdlib-jdk8"))
     implementation(Libs.freemarker)
 }
