+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/model/mapping/GeminioRecipeExpressionConverter.kt
@@ -8,14 +8,14 @@ import com.android.tools.idea.wizard.template.fragmentToLayout
 import com.android.tools.idea.wizard.template.layoutToActivity
 import com.android.tools.idea.wizard.template.layoutToFragment
 import com.android.tools.idea.wizard.template.underlinesToCamelCase
-import ru.hh.plugins.geminio.model.RecipeExpression
-import ru.hh.plugins.geminio.model.RecipeExpression.Command.*
+import ru.hh.plugins.geminio.sdk.model.recipe.RecipeExpression
+import ru.hh.plugins.geminio.sdk.model.recipe.RecipeExpression.Command.*
 import ru.hh.plugins.geminio.model.aliases.AndroidStudioTemplateBooleanParameter
 import ru.hh.plugins.geminio.model.aliases.AndroidStudioTemplateParameter
 import ru.hh.plugins.geminio.model.aliases.AndroidStudioTemplateParameterBooleanLambda
 import ru.hh.plugins.geminio.model.aliases.AndroidStudioTemplateParameterStringLambda
 import ru.hh.plugins.geminio.model.aliases.AndroidStudioTemplateStringParameter
-import ru.hh.plugins.geminio.model.enums.GeminioRecipeExpressionModifier.*
+import ru.hh.plugins.geminio.sdk.model.enums.GeminioRecipeExpressionModifier.*
 
 
 fun RecipeExpression.toBooleanLambda(
