+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/model/mapping/GeminioTemplateCategoryConverter.kt
@@ -1,7 +1,7 @@
 package ru.hh.plugins.geminio.model.mapping
 
 import com.android.tools.idea.wizard.template.Category
-import ru.hh.plugins.geminio.model.enums.GeminioTemplateCategory
+import ru.hh.plugins.geminio.sdk.model.enums.GeminioTemplateCategory
 
 
 /**
