+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/model/yaml/GeminioRecipeReader.kt
@@ -2,17 +2,17 @@ package ru.hh.plugins.geminio.model.yaml
 
 import ru.hh.plugins.extensions.EMPTY
 import ru.hh.plugins.geminio.GeminioConstants
-import ru.hh.plugins.geminio.model.GeminioRecipe
-import ru.hh.plugins.geminio.model.OptionalParams
-import ru.hh.plugins.geminio.model.RecipeCommand
-import ru.hh.plugins.geminio.model.RecipeExpression
-import ru.hh.plugins.geminio.model.RecipeParameter
-import ru.hh.plugins.geminio.model.RequiredParams
-import ru.hh.plugins.geminio.model.enums.GeminioStringParameterConstraint
-import ru.hh.plugins.geminio.model.enums.GeminioTemplateCategory
-import ru.hh.plugins.geminio.model.enums.GeminioTemplateConstraint
-import ru.hh.plugins.geminio.model.enums.GeminioTemplateFormFactor
-import ru.hh.plugins.geminio.model.enums.GeminioTemplateScreen
+import ru.hh.plugins.geminio.sdk.model.recipe.GeminioRecipe
+import ru.hh.plugins.geminio.sdk.model.recipe.OptionalParams
+import ru.hh.plugins.geminio.sdk.model.recipe.RecipeCommand
+import ru.hh.plugins.geminio.sdk.model.recipe.RecipeExpression
+import ru.hh.plugins.geminio.sdk.model.recipe.RecipeParameter
+import ru.hh.plugins.geminio.sdk.model.recipe.RequiredParams
+import ru.hh.plugins.geminio.sdk.model.enums.GeminioStringParameterConstraint
+import ru.hh.plugins.geminio.sdk.model.enums.GeminioTemplateCategory
+import ru.hh.plugins.geminio.sdk.model.enums.GeminioTemplateConstraint
+import ru.hh.plugins.geminio.sdk.model.enums.GeminioTemplateFormFactor
+import ru.hh.plugins.geminio.sdk.model.enums.GeminioTemplateScreen
 import ru.hh.plugins.geminio.model.yaml.YamlKeys.KEY_OPTIONAL_PARAMS
 import ru.hh.plugins.geminio.model.yaml.YamlKeys.KEY_OPTIONAL_PARAMS_CATEGORY
 import ru.hh.plugins.geminio.model.yaml.YamlKeys.KEY_OPTIONAL_PARAMS_CONSTRAINTS
