+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/template/executors/InstantiateCommandExecutor.kt
@@ -1,7 +1,7 @@
 package ru.hh.plugins.geminio.template.executors
 
 import com.android.tools.idea.wizard.template.RecipeExecutor
-import ru.hh.plugins.geminio.model.RecipeCommand
+import ru.hh.plugins.geminio.sdk.model.recipe.RecipeCommand
 import ru.hh.plugins.geminio.model.mapping.evaluateString
 import ru.hh.plugins.geminio.model.temp_data.GeminioRecipeExecutorData
 import java.io.File
