+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/template/geminioRecipe.kt
@@ -1,7 +1,7 @@
 package ru.hh.plugins.geminio.template
 
 import com.android.tools.idea.wizard.template.RecipeExecutor
-import ru.hh.plugins.geminio.model.GeminioRecipe
+import ru.hh.plugins.geminio.sdk.model.recipe.GeminioRecipe
 import ru.hh.plugins.geminio.model.temp_data.GeminioRecipeExecutorData
 import ru.hh.plugins.geminio.template.executors.executeCommands
